[Instrumentation] Move out to Utils (NFC)
Utility functions have been moved out to Utils.
