+++ llvm/lib/Transforms/Utils/Instrumentation.cpp
@@ -11,7 +11,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "llvm/Transforms/Instrumentation.h"
+#include "llvm/Transforms/Utils/Instrumentation.h"
 #include "llvm/IR/DiagnosticInfo.h"
 #include "llvm/IR/DiagnosticPrinter.h"
 #include "llvm/IR/IntrinsicInst.h"
@@ -41,7 +41,8 @@ bool llvm::checkIfAlreadyInstrumented(Module &M, StringRef Flag) {
 }
 
 /// Moves I before IP. Returns new insert point.
-static BasicBlock::iterator moveBeforeInsertPoint(BasicBlock::iterator I, BasicBlock::iterator IP) {
+static BasicBlock::iterator moveBeforeInsertPoint(BasicBlock::iterator I,
+                                                  BasicBlock::iterator IP) {
   // If I is IP, move the insert point down.
   if (I == IP) {
     ++IP;
@@ -93,7 +94,8 @@ GlobalVariable *llvm::createPrivateGlobalForString(Module &M, StringRef Str,
 }
 
 Comdat *llvm::getOrCreateFunctionComdat(Function &F, Triple &T) {
-  if (auto Comdat = F.getComdat()) return Comdat;
+  if (auto Comdat = F.getComdat())
+    return Comdat;
   assert(F.hasName());
   Module *M = F.getParent();
